More stategic changes
Now, staging and propulsion model parts have their own modules.
