+++ itsModelStaging.py
@@ -1,0 +1,143 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Jan 19 16:07:46 2018
+
+@author: carlos
+"""
+
+import numpy
+
+
+def stagingCalculate(con, Dv1: float, Dv2: float)-> None:
+
+    efflist = con['efflist']
+    Tlist = con['Tlist']
+
+    if con['homogeneous']:
+
+        if con['NStag'] == 0:
+
+            p2 = modelOptimalStagingHomogeneous([efflist[-1]],       Dv2,
+                                                [Tlist[-1]], con['Isp'],
+                                                con['g0'],  con['Mu'])
+            p1 = modelOptimalStagingHomogeneous([efflist[0]], Dv1 + Dv2,
+                                                [Tlist[0]], con['Isp'],
+                                                con['g0'],  con['Mu'])
+            p2.ms = p2.mu + p1.me[0]
+            p2.me[0] = 0.0
+            p1.NStag0or1(p2)
+
+        elif con['NStag'] == 1:
+
+            p2 = modelOptimalStagingHomogeneous([efflist[-1]],       Dv2,
+                                                [Tlist[-1]], con['Isp'],
+                                                con['g0'],  con['Mu'])
+            p1 = modelOptimalStagingHomogeneous([efflist[0]], Dv1 + Dv2,
+                                                [Tlist[0]], con['Isp'],
+                                                con['g0'],  con['Mu'])
+            p2.me[0] = p1.me[0]
+            p1.NStag0or1(p2)
+
+        else:
+
+            p2 = modelOptimalStagingHomogeneous([efflist[-1]], Dv2,
+                                                [Tlist[-1]], con['Isp'],
+                                                con['g0'],  con['Mu'])
+            p1 = modelOptimalStagingHomogeneous(efflist[0:-1], Dv1,
+                                                Tlist[0:-1], con['Isp'],
+                                                con['g0'], p2.mtot[0])
+
+    else:
+        raise Exception('itsme saying: heterogeneous vehicle is' +
+                        ' not supported yet!')
+
+    if p1.mtot[0]*con['g0'] > Tlist[0]:
+        raise Exception('itsme saying: weight greater than thrust!')
+
+    return p1, p2
+
+
+class modelOptimalStagingHomogeneous():
+    # optimalStaging() returns a object with information of
+    # optimal staging factor for a homogeneous vehcile
+    # The maximal staging reason is defined as reducing the total mass for
+    # a defined delta V.
+    # Structural eficience and thrust shall variate for diferent stages,  but
+    # specific impulse must be the same for all stages
+    # Based in Cornelisse (1979)
+
+    def __init__(self, effList: list, dV: float, Tlist: list, Isp: float,
+                 g0: float, mu: float):
+        self.Tlist = Tlist
+        self.e = numpy.array(effList)
+        self.T = numpy.array(Tlist)
+        self.dV = dV
+        self.mu = mu
+        self.Isp = Isp
+        self.c = Isp*g0
+        self.mflux = self.T/self.c
+
+        self._lamb = (numpy.exp(-self.dV/self.c/self.e.size) -
+                      self.e)/(1 - self.e)
+
+        phi = (1 - self.e)*(1 - self._lamb)
+
+        # Total sub-rocket mass
+        self.__mtotCalculate()
+        # Propelant mass on each stage
+        self.mp = self.mtot*phi
+        # Strutural mass of each stage
+        self.me = self.mp*(self.e/(1 - self.e))
+        # Duration of each stage burning
+        self.tb = self.mp/self.mflux
+        # Final burning time of each stage
+        self.__tfCalculate()
+        # Empty vehicle mass (!= 0.0 only for NStag == 0)
+        self.ms = 0.0
+        return None
+
+    def __mtotCalculate(self)-> None:
+
+        mtot = self.e*0.0
+        N = self.e.size-1
+        for ii in range(0,  N+1):
+            if ii == 0:
+                mtot[N - ii] = self.mu/self._lamb[N - ii]
+            else:
+                mtot[N - ii] = mtot[N - ii + 1]/self._lamb[N - ii]
+        self.mtot = mtot
+        return None
+
+    def __tfCalculate(self)-> None:
+
+        tf = self.e*0.0
+        N = self.tb.size-1
+        for ii in range(0, N+1):
+            if ii == 0:
+                tf[ii] = self.tb[ii]
+            else:
+                tf[ii] = self.tb[ii] + tf[ii-1]
+        self.tf = tf
+        return None
+
+    def NStag0or1(self, p: object)-> None:
+
+        self.dV = self.dV - p.dV
+        self.mu = p.mtot[0]
+        self.mp[0] = self.mp[0] - p.mp[0]  # Propelant mass on each stage
+        self.me[0] = self.me[0]*0.0
+        self.tb[0] = self.tb[0] - p.tb[0]  # Duration of each stage burning
+        self.tf[0] = self.tf[0] - p.tf[0]
+        return None
+
+    def printInfo(self):
+
+        print("\n\rdV =", self.dV)
+        print("mu =", self.mu)
+        print("me =", self.me)
+        print("mp =", self.mp)
+        print("mtot =", self.mtot)
+        print("mflux =", self.mflux)
+        print("tb =", self.tb)
+        print("tf =", self.tf)
